// packages/shared/src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      users: {
        Row: {
          avatar_url: string | null
          ban_expires_at: string | null
          ban_reason: string | null
          created_at: string
          id: string
          is_banned: boolean
          is_online: boolean
          last_seen: string | null
          updated_at: string
          username: string | null
          wallet_address: string
        }
        Insert: {
          avatar_url?: string | null
          ban_expires_at?: string | null
          ban_reason?: string | null
          created_at?: string
          id?: string
          is_banned?: boolean
          is_online?: boolean
          last_seen?: string | null
          updated_at?: string
          username?: string | null
          wallet_address: string
        }
        Update: {
          avatar_url?: string | null
          ban_expires_at?: string | null
          ban_reason?: string | null
          created_at?: string
          id?: string
          is_banned?: boolean
          is_online?: boolean
          last_seen?: string | null
          updated_at?: string
          username?: string | null
          wallet_address?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}